# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AutoFlow SaaS WhatsApp API
  version: 1.0.0
  description: API multiempresa para instancias, contatos, campanhas, mensagens e integracoes.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /instances:
    get:
      summary: Listar instancias da empresa
      responses:
        '200': {description: Lista de instancias}
    post:
      summary: Criar instancia respeitando o plano
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties: {name: {type: string, minLength: 2}}
      responses:
        '201': {description: Instancia criada}
        '409': {$ref: '#/components/responses/Error'}
  /instances/{id}/initialize:
    post:
      summary: Inicializar ou reconectar instancia
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses:
        '202': {description: Inicializacao solicitada}
  /instances/{id}/qr:
    get:
      summary: Obter QR temporario
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses:
        '200': {description: Estado e data URL do QR, se disponivel}
...
  /messages:
    post:
      security: [{apiKeyAuth: []}]
      summary: Enviar mensagem direta idempotente
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: {type: string, minLength: 8}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instanceId, text]
              oneOf:
                - required: [contactId]
                - required: [phoneNumber]
              properties:
                instanceId: {type: string, format: uuid}
                contactId: {type: string, format: uuid}
                phoneNumber: {type: string, pattern: '^[1-9][0-9]{9,14}$', description: 'Celular internacional de um contato existente. Informe contactId ou phoneNumber.'}
                text: {type: string, minLength: 1, maxLength: 4096}
      responses:
        '202': {description: Mensagem enfileirada}
        '409': {$ref: '#/components/responses/Error'}
  /messages/{id}:
    get:
      security: [{apiKeyAuth: []}]
      summary: Consultar estado de mensagem
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: Estado da mensagem}}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: http
      scheme: bearer
      description: Chave wa_live com escopo apropriado.
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: {type: string, format: uuid}
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: {type: string}
            message: {type: string}
        requestId: {type: string}
  responses:
...
    Error:
      description: Erro padronizado
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}